requirements.txt
backcall==0.2.0
certifi==2021.5.30
cycler==0.10.0
debugpy==1.4.1
importlib-metadata==3.10.1
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.1.0
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1
matplotlib-inline==0.1.2
numpy==1.21.2
opencv-python==4.5.3.56
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
prompt-toolkit==3.0.19
ptyprocess==0.7.0
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.2
PyYAML==5.4.1
pyzmq==22.1.0
scikit-learn==1.0
scipy==1.7.1
sklearn==0.0
threadpoolctl==3.0.0
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.2
wcwidth==0.2.5
zipp==3.5.0
